+++ src/frontend/apps/drive/src/features/auth/Auth.tsx
@@ -1,29 +1,16 @@
 import React, { PropsWithChildren, useEffect, useState } from "react";
 
-import { fetchAPI } from "@/features/api/fetchApi";
 import { User } from "@/features/auth/types";
-import { baseApiUrl } from "../api/utils";
-import { APIError } from "../api/APIError";
 import { posthog } from "posthog-js";
 import { SpinnerPage } from "@/features/ui/components/spinner/SpinnerPage";
-
-export const logout = () => {
-  window.location.replace(new URL("logout/", baseApiUrl()).href);
-  posthog.reset();
-};
-
-export const login = (returnTo?: string) => {
-  const url = new URL("authenticate/", baseApiUrl());
-  if (returnTo) {
-    url.searchParams.set("returnTo", returnTo);
-  }
-  window.location.replace(url.href);
-};
+import { getDriver } from "../config/Config";
 
 interface AuthContextInterface {
   user?: User | null;
   init?: () => Promise<User | null>;
   refreshUser?: () => Promise<void>;
+  login?: (returnTo?: string) => void;
+  logout?: () => void;
 }
 
 export const AuthContext = React.createContext<AuthContextInterface>({});
@@ -32,24 +19,17 @@ export const useAuth = () => React.useContext(AuthContext);
 
 export const Auth = ({
   children,
-  redirect,
-}: PropsWithChildren & { redirect?: boolean }) => {
+}: PropsWithChildren) => {
+  const driver = getDriver();
   const [user, setUser] = useState<User | null>();
 
   const init = async () => {
     try {
-      const response = await fetchAPI(`users/me/`, undefined, {
-        redirectOn40x: false,
-      });
-      const data = (await response.json()) as User;
+      const data = await driver.getConnectedUser();
       setUser(data);
       return data;
-    } catch (error) {
-      if (redirect && error instanceof APIError && error.code === 401) {
-        login();
-      } else {
-        setUser(null);
-      }
+    } catch {
+      setUser(null);
       return null;
     }
   };
@@ -57,6 +37,17 @@ export const Auth = ({
   const refreshUser = async () => {
     void init();
   };
+
+  const login = (returnTo?: string) => {
+    const loginUrl = driver.getLoginUrl(returnTo);
+    window.location.replace(loginUrl.href);
+  }
+
+  const logout = () => {
+    const logoutUrl = driver.getLogoutUrl();
+    window.location.replace(logoutUrl.href);
+    posthog.reset();
+  }
 
   useEffect(() => {
     void init();
@@ -80,6 +71,8 @@ export const Auth = ({
         user,
         init,
         refreshUser,
+        login,
+        logout
       }}
     >
       {children}